add get city and count hotel in the city

+++ src/main/java/com/juaracoding/tugasakhir/dto/respone/RespCityDTO.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 public class RespCityDTO {
 
     private String cityName;
-    private Integer totalHotel;
+    private Long totalHotel;
 }
